get_system {svDialogs}	R Documentation

Determine the appropriate GUI to instantiate

Description

The native version of the dialog box displays different versions depending on the OS (Linux, Windows or MacOS). In the case of 'RStudio', its own dialog boxes are used, if they exist. However, in 'RStudio Desktop' only, and when the argument rstudio = FALSE is set, OS dialog boxes may be forced, and the function would return the OS instead of RStudio.

Usage

get_system(rstudio = getOption("svDialogs.rstudio", TRUE))

Arguments

rstudio
Logical. Should 'RStudio' dialog boxes automatically be used if available? If FALSE, force using OS dialog boxes, but only in 'RStudio Desktop' (ignored in 'RStudio Server'). Can be changed globally with options(svDialogs.rstudio = TRUE|FALSE). TRUE by default.

Value

A character scalar giving either "RStudio" or the result of Sys.info()["sysname"]

[Package svDialogs version 1.1.0 Index]